Rename absolute import symbol from src to #
