+++ frontend/src/components/LoadMessageModal.tsx
@@ -1,10 +1,10 @@
 import React from "react";
 import { Button } from "@nextui-org/react";
-import { fetchMsgs, clearMsgs } from "src/services/session";
-import { sendChatMessageFromEvent } from "src/services/chatService";
-import { handleAssistantMessage } from "src/services/actions";
-import { ResFetchMsg } from "src/types/ResponseType";
-import toast from "src/utils/toast";
+import { fetchMsgs, clearMsgs } from "#/services/session";
+import { sendChatMessageFromEvent } from "#/services/chatService";
+import { handleAssistantMessage } from "#/services/actions";
+import { ResFetchMsg } from "#/types/ResponseType";
+import toast from "#/utils/toast";
 import ODModal from "./ODModal";
 
 interface LoadMessageModalProps {
