+++ frontend/src/components/file-explorer/TreeNode.test.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { act, render } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import TreeNode from "./TreeNode";
-import { WorkspaceFile } from "src/services/fileService";
+import { WorkspaceFile } from "#/services/fileService";
 
 const onFileClick = vi.fn();
