+++ frontend/src/components/file-explorer/TreeNode.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import FolderIcon from "../FolderIcon";
 import FileIcon from "../FileIcons";
-import { WorkspaceFile } from "src/services/fileService";
+import { WorkspaceFile } from "#/services/fileService";
 
 interface TitleProps {
   name: string;
