+++ frontend/src/components/settings/SettingsModal.test.tsx
@@ -11,7 +11,7 @@ import {
 import SettingsModal from "./SettingsModal";
 
 vi.mock("../../services/settingsService", async (importOriginal) => ({
-  ...(await importOriginal<typeof import("src/services/settingsService")>()),
+  ...(await importOriginal<typeof import("#/services/settingsService")>()),
   getCurrentSettings: vi.fn().mockReturnValue({}),
   saveSettings: vi.fn(),
   fetchModels: vi
