+++ frontend/src/components/settings/SettingsModal.tsx
@@ -8,9 +8,9 @@ import {
   fetchModels,
   getCurrentSettings,
   saveSettings,
-} from "src/services/settingsService";
-import { I18nKey } from "src/i18n/declaration";
-import { AvailableLanguages } from "src/i18n";
+} from "#/services/settingsService";
+import { I18nKey } from "#/i18n/declaration";
+import { AvailableLanguages } from "#/i18n";
 
 interface SettingsProps {
   isOpen: boolean;
