+++ frontend/src/services/settingsService.ts
@@ -1,9 +1,9 @@
-import { setInitialized } from "src/state/taskSlice";
-import store from "src/store";
-import ActionType from "src/types/ActionType";
-import { SupportedSettings } from "src/types/ConfigType";
-import { setByKey } from "src/state/settingsSlice";
-import toast from "src/utils/toast";
+import { setInitialized } from "#/state/taskSlice";
+import store from "#/store";
+import ActionType from "#/types/ActionType";
+import { SupportedSettings } from "#/types/ConfigType";
+import { setByKey } from "#/state/settingsSlice";
+import toast from "#/utils/toast";
 import Socket from "./socket";
 
 export type Settings = { [key: string]: string };
