+++ frontend/src/services/socket.ts
@@ -1,5 +1,5 @@
 // import { toast } from "sonner";
-import toast from "src/utils/toast";
+import toast from "#/utils/toast";
 import { handleAssistantMessage } from "./actions";
 import { getToken } from "./auth";
