+++ frontend/tsconfig.json
@@ -22,7 +22,7 @@
     "types": ["vite/client", "vitest/globals"],
     "baseUrl": ".",
     "paths": {
-      "src/*": ["./src/*"]
+      "#/*": ["./src/*"]
     }
   },
   "include": [
